Build templates feature: copy files and replace tokens without configuration
This feature allows you to automatically apply templates (replacing DATE, VERSION, PRO tokens in files) without extra configuration. You only have to use the correct directory structure in /build/templates and remove your current copy/replacetokens code fragments of the extension build.xml files.

This will save you a lot of configuration. The only downside is that it is NOT backward compatible with the templates directory as explained above. However though I’ve used this for quite a while and I *really* love it and I can really recommend it.